delay axis assurance until after batching-assessments
